search page added. Modified schema; changed amount fields to float. DB reset required

+++ db/migrate/20151105081922_create_verification_requests.rb
@@ -7,8 +7,8 @@ class CreateVerificationRequests < ActiveRecord::Migration
       t.string :name
       t.string :hallticket_no
       t.string :document_link
-      t.integer :amount
-      t.integer :service_tax
+      t.float :amount
+      t.float :service_tax
